EF Core: Create the database using migrations
On the .NET Core CLI:
```
dotnet tool install --global dotnet-ef
dotnet add package Microsoft.EntityFrameworkCore.Design
dotnet ef migrations add InitialCreate
dotnet ef database update
```
This installs dotnet ef and the design package which is required to run
the command on a project. The migrations command scaffolds a migration
to create the initial set of tables for the model. The database update
command creates the database and applies the new migration to it.

Using Visual Studio:
```
Install-Package Microsoft.EntityFrameworkCore.Tools
Add-Migration InitialCreate
Update-Database
```

This installs the PMC tools for EF Core. The Add-Migration command
scaffolds a migration to create the initial set of tables for the model.
The Update-Database command creates the database and applies the new
migration to it.